requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-cors-headers==3.13.0
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-import-export==3.0.2
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.7.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
flower==1.2.0
gevent==22.10.2
greenlet==2.0.1
humanize==4.4.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tablib==3.3.0
text-unidecode==1.3
tornado==6.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.6
zope.interface==5.5.2
